improve: GpCommandLineParser: allow to restrict for assign published properties only
Sometimes a target class can have a public property of a custom
enumeration type which is set from a string value from a command line.
E.g.
  type
    TLogLevel = (llTrace, llDebug, ...

    TCommandLine = class
    private
      FLogLevel: TLogLevel
    public
      property LogLevel: TLogLevel read FLogLevel write FLogLevel;

      [CLPLongName('log-level')]
      property LogLevelAsStr: string read GetLogLevelAsStr write SetLogLevelAsStr;
    end;

In this case TGpCommandLineParser raises the SUnsupportedPropertyType
exception for LogLevel property as it supports Integer, String and
Boolean types only.

Let's fix it by adding opPublishedPropertiesOnly option to tell a parser
to respect published properties only.
So, TCommandLine will appear to be defined as the following

    TCommandLine = class
    private
      FLogLevel: TLogLevel
    public
      property LogLevel: TLogLevel read FLogLevel write FLogLevel;
    published
      [CLPLongName('log-level')]
      property LogLevelAsStr: string read GetLogLevelAsStr write SetLogLevelAsStr;
    end;